Initial working test (finally)
